moved the jumbotron data from App.js to Jumbotron.js
